Add some new specs and refactor warchest file
