+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require "bundler/setup"
-require "warchest_lite"
+require "warchest"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
