--- spec/warchest_lite_spec.rb
@@ -1,16 +1,0 @@
-RSpec.describe WarchestLite do
-  describe Game do
-    let(:player1) {Player.new('Player_1')}
-    let(:player2) {Player.new('Player_2')}
-    describe '#create_board' do
-
-    end
-  end
-  it "has a version number" do
-    expect(WarchestLite::VERSION).not_to be nil
-  end
-
-  it "does something useful" do
-    expect(false).to eq(true)
-  end
-end
